+++ packages/compilers/src/lib/solidityCompiler.ts
@@ -71,7 +71,7 @@ export async function useSolidityCompiler(
   }
   let startCompilation: number;
   if (solcPath && !forceEmscripten) {
-    logInfo('Compiling with solc binary', { version, solcPath });
+    logDebug('Compiling with solc binary', { version, solcPath });
     startCompilation = Date.now();
     try {
       compiled = await asyncExec(
@@ -86,7 +86,7 @@ export async function useSolidityCompiler(
       throw error;
     }
   } else {
-    logInfo('Compiling with solc-js', { version });
+    logDebug('Compiling with solc-js', { version });
     const solJson = await getSolcJs(solJsonRepoPath, version);
     startCompilation = Date.now();
     if (solJson) {
@@ -115,7 +115,7 @@ export async function useSolidityCompiler(
   }
 
   const endCompilation = Date.now();
-  logInfo('Local compiler - Compilation done', {
+  logDebug('Local compiler - Compilation done', {
     compiler: 'solidity',
     timeInMs: endCompilation - startCompilation,
   });