+++ services/monitor/test/Monitor.spec.ts
@@ -170,6 +170,8 @@ describe("Monitor", function () {
   });
 
   it("should throw an error if there are chainConfigs for chains not being monitored", () => {
+    const nodeEnv = process.env.NODE_ENV;
+    process.env.NODE_ENV = "production";
     expect(
       () =>
         new Monitor([localChain], {
@@ -180,6 +182,24 @@ describe("Monitor", function () {
     ).to.throw(
       "Chain configs found for chains that are not being monitored: 2",
     );
+    process.env.NODE_ENV = nodeEnv;
+  });
+
+  it("should enable factory monitoring when MONITOR_FACTORIES env var is set to true", () => {
+    const monitorFactories = process.env.MONITOR_FACTORIES;
+    process.env.MONITOR_FACTORIES = "true";
+    const monitor = new Monitor([localChain]);
+
+    expect(monitor["config"].monitorFactories).to.be.true;
+    monitor["chainMonitors"].forEach((chainMonitor) => {
+      expect(chainMonitor["monitorFactories"]).to.be.true;
+    });
+
+    if (monitorFactories !== undefined) {
+      process.env.MONITOR_FACTORIES = monitorFactories;
+    } else {
+      delete process.env.MONITOR_FACTORIES;
+    }
   });
 
   it("should successfully catch a deployed contract, assemble, and send to Sourcify", async () => {